repository added and Int to Long

+++ app/src/main/java/com/example/mcdonaldsclone/core/database/local/Entities/CartItemEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "cart_items")
 data class CartItemEntity(
-    @PrimaryKey val id: Int,
-    val productId: Int,
+    @PrimaryKey val id: Long,
+    val productId: Long,
     val quantity: Int
 )
