+++ app/src/main/java/com/example/mcdonaldsclone/core/database/local/Entities/UserSettingsEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "user_settings")
 data class UserSettingsEntity(
-    @PrimaryKey val id: Int = 0,
+    @PrimaryKey val id: Long = 0,
     val isDarkMode: Boolean,
     val preferredLanguage: String
 )
